+++ tests/Unit/Public/Test-M365DSCPowerShellDataFile.tests.ps1
@@ -47,6 +47,7 @@ Describe Test-M365DSCPowershellDataFile {
                                 UniqueId = 'String | Required | Required String'
                                 Item3 = 'Boolean | Required | Required Boolean'
                                 Item4 = 'Guid | Optional | Optional Guid'
+                                Item5 = 'String | Required | Required String'
                             }
                         )
                     }
@@ -56,7 +57,7 @@ Describe Test-M365DSCPowershellDataFile {
     }
 
     Context 'Test function Test-M365DSCPowershellDataFile' {
-        It 'Should successfully test of the inputted object' {
+        It 'Should successfully test of the inputted object - Testing type and value' {
             $Object = @{
                 NonNodeData = @{
                     'Exchange' = @{
@@ -82,11 +83,11 @@ Describe Test-M365DSCPowershellDataFile {
                 }
             }
 
-            $result = Test-M365DSCPowershellDataFile -Test TypeValue -InputObject $Object -pesterVerbosity None -pesterOutputObject
+            $result = Test-M365DSCPowershellDataFile -Test TypeValue -InputObject $Object -PesterVerbosity None -PesterOutputObject
             $result.Result | Should -Be 'Passed'
         }
 
-        It 'Should fail test of the inputted object' {
+        It 'Should fail test of the inputted object - Testing type and value' {
             $Object = @{
                 NonNodeData = @{
                     'Exchange' = @{
@@ -111,7 +112,244 @@ Describe Test-M365DSCPowershellDataFile {
                 }
             }
 
-            $result = Test-M365DSCPowershellDataFile -Test TypeValue -InputObject $Object -pesterVerbosity None -pesterOutputObject
+            $result = Test-M365DSCPowershellDataFile -Test TypeValue -InputObject $Object -PesterVerbosity None -PesterOutputObject
+            $result.Result | Should -Be 'Failed'
+        }
+        It 'Should successfully test of the inputted object - Testing Required' {
+            $Object = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item1 = 'String'
+                            Item2 = 5
+                        }
+                    }
+                    'Teams' = @{
+                        Resource2 = @(
+                            @{
+                                UniqueId = "Test"
+                                Item3 = $true
+                                Item4 = 'c25d6579-be90-4484-81ef-9280d4817440'
+                            }
+                            @{
+                                UniqueId = "Test2"
+                                Item3 = $false
+                                Item4 = 'c25d6579-be90-4484-81ef-9280d4817441'
+                                Item5 = 'String'
+                            }
+                        )
+                    }
+                }
+            }
+
+            $result = Test-M365DSCPowershellDataFile -Test Required -InputObject $Object -PesterVerbosity None -PesterOutputObject
+            $result.Result | Should -Be 'Passed'
+        }
+
+        It 'Should fail test of the inputted object - Testing Required' {
+            $Object = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item1 = 'String'
+                            Item2 = 5
+                        }
+                    }
+                    'Teams' = @{
+                        Resource2 = @(
+                            @{
+                                UniqueId = "Test"
+                                Item3 = $true
+                                Item4 = 'c25d6579-be90-4484-81ef-9280d4817440'
+                            }
+                            @{
+                                UniqueId = "Test2"
+                                ItemUnknown = $false
+                            }
+                        )
+                    }
+                }
+            }
+
+            $result = Test-M365DSCPowershellDataFile -Test Required -InputObject $Object -pesterVerbosity None -pesterOutputObject
+            $result.Result | Should -Be 'Failed'
+        }
+        It 'Should successfully test of the inputted object - Testing Mandatory (Present)' {
+            $Object = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item1 = 'String'
+                            Item2 = 5
+                        }
+                    }
+                    'Teams' = @{
+                        Resource2 = @(
+                            @{
+                                UniqueId = "Test"
+                                Item3 = $true
+                                Item4 = 'c25d6579-be90-4484-81ef-9280d4817440'
+                            }
+                            @{
+                                UniqueId = "Test2"
+                                Item3 = $false
+                                Item4 = 'c25d6579-be90-4484-81ef-9280d4817441'
+                                Item5 = 'String'
+                            }
+                        )
+                    }
+                }
+            }
+            $MandatoryObject = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item1 = 'String'
+                        }
+                    }
+                    'Teams' = @{
+                        Resource2 = @(
+                            @{
+                                UniqueId = "Test"
+                                Item3 = $true
+                            }
+                            @{
+                                UniqueId = "Test2"
+                                Item5 = 'String'
+                            }
+                        )
+                    }
+                }
+            }
+
+            $result = Test-M365DSCPowershellDataFile -Test Mandatory -InputObject $Object -MandatoryObject $MandatoryObject -MandatoryAction 'Present' -PesterVerbosity None -PesterOutputObject
+            $result.Result | Should -Be 'Passed'
+        }
+
+        It 'Should fail test of the inputted object - Testing Mandatory (Present)' {
+            $Object = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item1 = 'String'
+                            Item2 = 5
+                        }
+                    }
+                    'Teams' = @{
+                        Resource2 = @(
+                            @{
+                                UniqueId = "Test"
+                                Item3 = $true
+                                Item4 = 'c25d6579-be90-4484-81ef-9280d4817440'
+                            }
+                            @{
+                                UniqueId = "Test2"
+                                ItemUnknown = $false
+                            }
+                        )
+                    }
+                }
+            }
+            $MandatoryObject = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item1 = 'String'
+                        }
+                    }
+                    'Teams' = @{
+                        Resource2 = @(
+                            @{
+                                UniqueId = "Test"
+                                Item3 = $false
+                            }
+                            @{
+                                UniqueId = "Test2"
+                                Item5 = 'StringWrong'
+                            }
+                        )
+                    }
+                }
+            }
+
+            $result = Test-M365DSCPowershellDataFile -Test Mandatory -InputObject $Object -MandatoryObject $MandatoryObject -MandatoryAction 'Present' -PesterVerbosity None -PesterOutputObject
+            $result.Result | Should -Be 'Failed'
+        }
+        It 'Should successfully test of the inputted object - Testing Mandatory (Absent)' {
+            $Object = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item1 = 'String'
+                            Item2 = 5
+                        }
+                    }
+                    'Teams' = @{
+                        Resource2 = @(
+                            @{
+                                UniqueId = "Test"
+                                Item3 = $true
+                                Item4 = 'c25d6579-be90-4484-81ef-9280d4817440'
+                            }
+                            @{
+                                UniqueId = "Test2"
+                                Item3 = $false
+                                Item4 = 'c25d6579-be90-4484-81ef-9280d4817441'
+                                Item5 = 'String'
+                            }
+                        )
+                    }
+                }
+            }
+            $MandatoryObject = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item3 = 'String'
+                        }
+                    }
+                }
+            }
+
+            $result = Test-M365DSCPowershellDataFile -Test Mandatory -InputObject $Object -MandatoryObject $MandatoryObject -MandatoryAction 'Absent' -PesterVerbosity None -PesterOutputObject
+            $result.Result | Should -Be 'Passed'
+        }
+
+        It 'Should fail test of the inputted object - Testing Mandatory (Absent)' {
+            $Object = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item1 = 'String'
+                            Item2 = 5
+                        }
+                    }
+                    'Teams' = @{
+                        Resource2 = @(
+                            @{
+                                UniqueId = "Test"
+                                Item3 = $true
+                                Item4 = 'c25d6579-be90-4484-81ef-9280d4817440'
+                            }
+                            @{
+                                UniqueId = "Test2"
+                                ItemUnknown = $false
+                            }
+                        )
+                    }
+                }
+            }
+            $MandatoryObject = @{
+                NonNodeData = @{
+                    'Exchange' = @{
+                        Resource1 = @{
+                            Item1 = 'String'
+                        }
+                    }
+                }
+            }
+
+            $result = Test-M365DSCPowershellDataFile -Test Mandatory -InputObject $Object -MandatoryObject $MandatoryObject -MandatoryAction 'Absent' -PesterVerbosity None -PesterOutputObject
             $result.Result | Should -Be 'Failed'
         }
     }